feat: add electron support
