+++ daemon/src/identify/protocol.rs
@@ -115,6 +115,7 @@ pub enum Environment {
     MyNode,
     Docker,
     Binary,
+    Electron,
     Test,
     Legacy,
     Unknown,
@@ -130,6 +131,7 @@ impl From<crate::Environment> for Environment {
             crate::Environment::MyNode => Environment::MyNode,
             crate::Environment::Docker => Environment::Docker,
             crate::Environment::Binary => Environment::Binary,
+            crate::Environment::Electron => Environment::Electron,
             crate::Environment::Test => Environment::Test,
             crate::Environment::Legacy => Environment::Legacy,
             crate::Environment::Unknown => Environment::Unknown,
@@ -147,6 +149,7 @@ impl From<Environment> for crate::Environment {
             Environment::MyNode => crate::Environment::MyNode,
             Environment::Docker => crate::Environment::Docker,
             Environment::Binary => crate::Environment::Binary,
+            Environment::Electron => crate::Environment::Electron,
             Environment::Test => crate::Environment::Test,
             Environment::Legacy => crate::Environment::Legacy,
             Environment::Unknown => crate::Environment::Unknown,
