+++ taker-electron/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "itchysats"
+version = "0.1.0"
+edition = "2021"
+exclude = ["index.node"]
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+neon = { version = "0.10.1", default-features = false, features = ["channel-api", "napi-6", "promise-api"] }
+once_cell = "1"
+taker = { path = "../taker" }
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "net", "tracing"] }
