+++ taker-electron/src/index.d.ts
@@ -1,0 +1,9 @@
+export {};
+
+declare global {
+    interface Window {
+        itchysats: {
+            load: () => void;
+        };
+    }
+}
